Fix(imports): resolve module-level import not at top of file linting errors
- Move 'from typing import Any as _Any' to top of step_policies.py
- Move import statement before docstring in state_machine_policy.py
- Fixes E402 linting errors in both files
- Maintains functionality while ensuring proper import organization
- All linting checks now pass
